sencente + translation add backend

+++ LanguageApp/src/main/java/proj/LanguageApp/Repository/WordRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import proj.LanguageApp.Entity.Word;
 
-public interface WordRepository extends JpaRepository<Word, Integer> {
+public interface WordRepository extends JpaRepository<Word, Long> {
 
     @Modifying
     @Query("delete from Word where id = :id")
